updated functions
When testing a blog post, I came across an error when the script tried to scrape it. One of my previous fixes hadn't taken into account that not all block elements marked with the CSS id of "secondary" in WordPress are <div>s. For this blog post, it was an <aside>. So I added this in as an if statement.

The second amendment was in checks for entities within the existing_tags list. They didn't always return a 1 when they should have for all caps strings (e.g. API) so that was included as a check. I also found a way to check if any of the iterables in the existing_tags list were in the entity iterable via an any() comprehension (thanks Stack Overflow) and after some tweaking, it worked. It's not perfect though. As an example, one of my tags is "programming" and that string would feature in entities such as "Python (programming language)" which is fine because they're semantically related but they aren't the exact same entity. It was a choice of keeping it as it was and missing existing tags that matched or a fuzzy search result and I chose the latter.